Refactoring Crud ACL

+++ app/Http/Controllers/Acl/RoleController.php
@@ -10,7 +10,6 @@ use Illuminate\Support\Facades\Gate;
 class RoleController extends Controller
 {
     private $repository;
-    private $permission;
 
     public function __construct(RoleRepositoryInterface $repository, Request $request)
     {
@@ -19,21 +18,21 @@ class RoleController extends Controller
 
     public function search(Request $request)
     {   
-       // Gate::authorize($this->permission);
+        Gate::authorize('search_role');
 
         return $this->repository->search($request);
     }
 
     public function store(RoleRequest $request)
     {   
-       // Gate::authorize($this->permission);
+        Gate::authorize('store_role');
 
         return $this->repository->store($request);
     }
 
     public function show($id)
     {   
-       // Gate::authorize($this->permission);
+        Gate::authorize('show_role');
 
         return $this->repository->show($id);
     }
@@ -41,42 +40,42 @@ class RoleController extends Controller
 
     public function update(RoleRequest $request)
     {   
-       // Gate::authorize($this->permission);
+        Gate::authorize('update_role');
 
         return $this->repository->update($request);
     }
 
     public function activate($id)
     {   
-       // Gate::authorize($this->permission);
+        Gate::authorize('activate_role');
 
         return $this->repository->activate($id);
     }
 
     public function inactivate($id)
     {   
-       // Gate::authorize($this->permission);
+        Gate::authorize('inactivate_role');
 
         return $this->repository->inactivate($id);
     }
 
     public function deleted($id)
     {   
-       // Gate::authorize($this->permission);
+        Gate::authorize('deleted_role');
 
         return $this->repository->deleted($id);
     }
 
     public function recover($id)
     {   
-       // Gate::authorize($this->permission);
+        Gate::authorize('recover_role');
 
         return $this->repository->recover($id);
     }
     
     public function destroy($id)
     {   
-       // Gate::authorize($this->permission);
+        Gate::authorize('destroy_role');
 
         return $this->repository->destroy($id);
     }
